// src/api/Api.ts
export interface ExtracurricularField {
  /**
   * 비교과 이름
   * @example "A비교과"
   */
  originTitle?: string;
  /**
   * 비교과 url
   * @example "https://abc.com"
   */
  url?: string;
  /**
   * 신청시작 일자
   * @format date-time
   */
  applicationStart?: string;
  /**
   * 신청종료 일자
   * @format date-time
   */
  applicationEnd?: string;
  /**
   * 활동시작 일자
   * @format date-time
   */
  activityStart?: string;
  /**
   * 활동종료 일자
   * @format date-time
   */
  activityEnd?: string;
}

export interface GetScheduleDetailResponse {
  /**
   * 스케쥴 제목
   * @example "A비교과"
   */
  title?: string;
  /**
   * 스케쥴 상세정보
   * @example "공C487에서 열릴 예정, 시간은 17시 ~ 19시"
   */
  content?: string;
  /**
   * 스케쥴 타입
   * @example "EXTRACURRICULAR(비교과 관련), NORMAL(일반 일정)"
   */
  scheduleType?: "EXTRACURRICULAR" | "NORMAL";
  /**
   * 시작 날짜
   * @format date-time
   */
  startDateTime?: string;
  /**
   * 끝 날짜
   * @format date-time
   */
  endDateTime?: string;
  /** 관련된 비교과(일반 일정일 경우에는 null값이 들어옴) */
  extracurricularField?: ExtracurricularField;
}